btrfs-progs: move send-stream.c to common/
