+++ common/send-stream.c
@@ -20,7 +20,7 @@
 #include <unistd.h>
 
 #include "send.h"
-#include "send-stream.h"
+#include "common/send-stream.h"
 #include "crypto/crc32c.h"
 #include "common/utils.h"
